feat: 이용자 등급 구분 추가

+++ src/products/service.ts
@@ -1,19 +1,35 @@
 import repository from "./repository";
+import { isAdmin } from "./constant/authority"
 
 export default {
   getAll: async () => {
     return await repository.getAll()
   },
-  post: async (body: any) => {
-    return await repository.post(body)
+  post: async (body: any, headers: any) => {
+    const Grade = headers.grade
+    if (isAdmin(Grade)) {
+      return await repository.post(body)
+    } else {
+      throw Error()
+    }
   },
   getOne: async (productId: number) => {
     return await repository.getOne(productId)
   },
-  patch: async (productId: number, body: any) => {
-    return await repository.patch(productId, body)
+  patch: async (productId: number, body: any, headers: any) => {
+    const Grade = headers.grade
+    if (isAdmin(Grade)) {
+      return await repository.patch(productId, body)
+    } else {
+      throw Error()
+    }
   },
-  delete: async (productId: number) => {
-    return await repository.delete(productId)
+  delete: async (productId: number, headers: any) => {
+    const Grade = headers.grade
+    if (isAdmin(Grade)) {
+      return await repository.delete(productId)
+    } else {
+      throw Error()
+    }
   },
 }
